+++ merged-packages/utils/src/caip-types.ts
@@ -148,8 +148,12 @@ export enum KnownCaipNamespace {
  *
  * This is a narrower, more type-safe alternative to {@link CaipChainId} for use cases
  * where the chain namespace must be one of the known standards.
+ *
+ * @template Namespace - The namespace of the CAIP-2 chain ID. Must be a known namespace specified in {@link KnownCaipNamespace}.
  */
-export type KnownCaipNamespacedChainId = `${KnownCaipNamespace}:${string}`;
+export type KnownCaipNamespacedChainId<
+  Namespace extends `${KnownCaipNamespace}` = `${KnownCaipNamespace}`,
+> = `${Namespace}:${string}`;
 
 /**
  * Check if the given value is a {@link CaipChainId}.